Appdelegateからコードを分割
