+++ QOLAB/QOLAB/Monitor/CallBackFunctions.swift
@@ -136,7 +136,7 @@ Outside:if pressed == 1
             if CallBackFunctions.CAPSLOCK
             {
                 if scancode >= 0 && scancode <= 100 {
-                    let d = AppDelegate()
+                    let d = Sensing()
                     d.keyCountUp(key: mySelf.keyMap[scancode]![1])
                     d.keyCountUpForSitting(key: mySelf.keyMap[scancode]![1])
                 }
@@ -145,7 +145,7 @@ Outside:if pressed == 1
             {
                 if scancode >= 0 && scancode <= 100 {
                     print("else:", scancode)
-                    let d = AppDelegate()
+                    let d = Sensing()
                     d.keyCountUp(key: mySelf.keyMap[scancode]![0])
                     d.keyCountUpForSitting(key: mySelf.keyMap[scancode]![0])
                 }
